Update the FindMKL to account for MKL installed by itself
The previous FindMKL relied upon the CMAKE_Fortran_COMPILER to find the
various libraries. Since BCLS does not require a Fortran compiler
this is a bit of overkill and will not work in the general case. This fix
should allow BCLS to be built with just the MKL library installed.

Tested on macOS only.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>